build: fix iconv link issue with non Linux systems
This is a hacky workaround as the m4 scripts for iconv seem to set
LIBICONV to '-liconv' despite that this lib is not available if iconv
functionality is coming out of libc. But we need -liconv for other
systems, such as Mac OS. Other projects I looked seem to have their own
set of iconv m4 macros... why?